Functional command handlers for LIN and PTP
Right now they're just tossed in Driver.  I need to pull them out into some other interface so that they can be registered for any type.  I'm not sure if I want to make a command list for each robot type or try to figure out what joints and aux joints are needed.  That could be very difficult since Keba's robots are made by the customer and can be in a ton of different configurations.