Async kinda works
This is not ready for actual use.  After fighting with asio, I finally got it to work.  Every change I make seems to break it, so I'm committing before refactoring.

Aborting is possible.  It can kill a command right in the middle of the send/receive cycle.

The key things were:  
-Do NOT put a .join on a thread launched in an async_connect.  I wanted to do this to automatically reconnect.  This will NOT work.  Maybe I can just call the connect function again from the thread before it ends.
- DO make a boost::asio::io_service::work work_ and a thread for io_service.run().