Attempt to deal with out-of-order messages after a socket cancel
It was possible to get an out of order "done" after cancelling a command since it was just cancelling all async operations.  If the PLC responded, this would leave a "done" message sitting in the socket, just waiting for the next read.  That would make the read immediately complete with the old message.  

I want ABORT to be able to deal with both normal stops and a stuck PLC command.  What I'm doing  when I get an ABORT is:
- Use async to launch cancelSocketCmd with a timeout of 50ms
   - It will try to get the cmd socket mutex for timeout. If it gets the mutex, then the last Cmd was already responded to and won't block anything.  If it was called with replace_commands (which it should be) the cmd lost will be empty.
- sendCommand to abort command via the Get channel
- Wait and hope the PLC responds before the timeout expires
   - If it does, great, nothing more to do
   - If not, call cancel on the cmd socket and cmdSocketFlusher.  It will continue to run until flush_socket_cmd_ is false and the socket is cancelled.