Start moving command_list handler into Connection
This is needed to support multiple robots.  I'm going to use the namespace of the command_list to signal which robot to use.  I started doing this in Driver, but it didn't make any sense.  I have not removed the old callback from the driver yet.  

I may make a new Robot class to sit in between Driver and Connection.  I originally didn't want any ros message stuff in the Connection.