Give each CommandHandler knowledge of its CommandRegister
Basic: 
- CommandHandler::registerCommands is protected.  
- CommandRegister::initialize must be called with the joint list vector.  It will call registerCommands.
- CommandHandlers can only be added via CommandRegister::addHandler, which will also set a pointer back to the CommandRegister.
- Overloaded CommandHandler classes should implement getCommandRegister and cast a return to their type.

It has become necessary for the CommandHandlers to have more information about the robot/joint configuration.  I don't want to have to make a ton of constructors/setters and set every piece of data that gets added to the CommandRegister in each handler.  

The easiest thing to do is give each CommandHandler a pointer back to the CommandRegister that made it.  I didn't want to have to manually set it each time.  The CommandRegister::addHandler methods will take care of this.  

The templated overload of addHandler handles things pretty cleanly.  I don't like the one for CommandHandler&& since you have to remember to std::move.