Don't clear command_list on eof error
I ran into this a lot while writing server_test.py, which involved restarting it a lot.  The Get socket will reconnect immediately since it's always sending stuff, but Cmd will sit there until it gets a command, find out the socket was closed on the server, then have to restart itself.  

Preserve the list but allow the CmdThread to restart so it can be sent after reconnecting.  If reconnecting fails, the command_list will always be cleared.