Publish tool_frame_pose as PoseStamped
Convert the tool_frame results into a PoseStamped.  This is the robot's reported position with orientation as a quaternion.  This could be monitored to make sure the correct tool is active, the base is set correctly, etc.  

It's still a mess.  I need to move the pose calculation somewhere nicer than publishState().