Close all Connector connections before stopping driver
The Driver will call stop() on each connection during shutdown.  This closes the sockets and joins the threads.

This seems to fix the hanging during a ctrl+c on the PLC.  I never had a problem while running it on my pc, but the more resource constrained PLC running an older version of Debian would hang.