Handle disconnects better for Get thead
Use future::wait_for(500ms) for Get commands and close the socket if it fails.  Cmds can delay forever and the next time you try to send a command to a disconnected socket it will error out and be handled.  Get commands don't overlap so if you pull a cable it will never disconnect.